% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysoundex.R
\name{my_soundex}
\alias{my_soundex}
\title{My soundex}
\usage{
my_soundex(strings, useBytes = FALSE)
}
\arguments{
\item{strings}{\code{[character]} vector. Strings to encode to soundex.}

\item{useBytes}{\code{[logical]} scalar. Should the strings be treated
as a byte sequence or by \code{UTF} code point?}
}
\description{
A soundex encoder with a message.
}
\section{Details}{

\code{strings} is coerced to \code{character} using \code{\link[base]{as.character}}
If \code{useBytes==FALSE}, the input string is converted to \code{UTF-8} prior
to soundex encoding (this has no effect on \code{ASCII} strings).
}

\section{References}{


See \code{\link[stringdist]{stringdist}} for details on the soundex algorithm.
}

\examples{

names <- c("Hubert Farnsworth","Bender Bending Rodriquez")
my_soundex(names)

}
